@@ -65,12 +65,12 @@
     "tweezer.js": "^1.4.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
+    "@babel/core": "^7.16.0",
+    "@babel/preset-env": "^7.16.0",
     "autoprefixer-stylus": "^1.0.0",
     "axios": "^0.21.1",
     "babel-eslint": "^10.0.3",
-    "babel-jest": "^26.3.0",
+    "babel-jest": "^27.0.0",
     "browser-sync": "^2.26.12",
     "chokidar": "^3.4.2",
     "common-tags": "^1.8.0",
@@ -78,13 +78,14 @@
     "copy-dir": "^1.2.0",
     "cross-env": "^6.0.3",
     "cssnano": "^4.1.10",
-    "eslint": "^5.16.0",
+    "es-main": "^1.0.2",
+    "eslint": "^7.0.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jest": "^24.0.2",
     "eslint-plugin-jest-playwright": "^0.2.1",
     "eslint-plugin-prettier": "^3.1.2",
     "husky": "^3.1.0",
-    "jest": "^26.4.2",
+    "jest": "^27.0.0",
     "jest-image-snapshot": "^4.2.0",
     "jest-playwright-preset": "^1.4.1",
     "lerna": "^3.22.1",
