@@ -1,4 +1,6 @@
-import mock from 'xhr-mock';
+import _mock from 'xhr-mock/lib/index.js';
+
+const mock = _mock.default;
 
 const sideEffects = {
   document: {
